Fixed a bug and satisfied a warning
There was a bug in tests-tower.cpp that was unrelated to the
implementation. An int conversion warning had popped up in program.cpp
so I static_casted the size_t to int.

+++ src/program.cpp
@@ -52,7 +52,8 @@ void Program::calc_above_weight()
 void Program::check_balance()
 {
     if (children.size() == 0 ||
-            above_weight == children.size() * children.front()->get_weight())
+            above_weight == static_cast<int>(children.size()) *
+                children.front()->get_weight())
         balanced = true;
     else
         balanced = false;
